Implement masking of secrets in internal logger. Also enforce masking in colorized log if a Printf used
This commit implements the masking of secrets in logger output and revises how debugging is set up, with a focus on logs involving secrets. The secrets are made available only after the playbook is loaded, and they're now passed as parameters to setupLog function. Relevant tests were updated and an additional test case for Printf function with secrets was added.