Add system env variable expansion to envVars function (#186)
* Add system env variable expansion to envVars function

The function `envVars` has been updated to enable the application to also read system environment variables, in addition to CLI and env file variables. This change includes an update of the function's comment to reflect the new behavior, and the addition of new test cases to validate the feature. CLI variables override env file variables, which in turn override system environment variables if they share the same name.

* add a test for env file with expanded os env

* add docs about os env expansion