Add second verbosity level and update related logic
In this update, a second verbosity level (Verbose2) has been added to the codebase. The normal -v (Verbose) is not reporting the content of the script anymore but -vv (Verbose2) only. Additionally, the verbosity short flag in cmd/spot/main.go has been modified and its type to boolean slice to allow increased verbosity when multiple "v" flags are used.