Fix incorrect handling of task-level user #192
The change eliminates setting the default task.User on the config level,
and moves this responsibility to the runner. Now the runner explicitly
checks the presence of task.User and passes it in.

In addition, runTaskOnHost sets the passed user into activeTask
(a copy of the task used later on). This was done to be able to know what
the actual user is. Currently, the only use is in setting SPOT_REMOTE_USER,
however, it can also be used in any place that needs to know the active remote user.