Collect and process register variables in the runner
The runner now differentiates between all variables and registered variables from command execution. Registered variables are a subset of all variables that will be used for subsequent tasks. This additional feature aids in more specific tracking and utilization of key variables in more complex scripting scenarios. This is a preparation step to implement cross-task registered vars #196